requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
diskcache==5.6.3
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
llama_cpp_python==0.3.8
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
synap-python @ https://github.com/synaptics-synap/synap-python/releases/download/v0.0.3-preview/synap_python-0.0.3-cp310-cp310-linux_aarch64.whl
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
